+++ src/shillelagh/adapters/api/gsheets/exceptions.py
@@ -1,0 +1,11 @@
+"""
+Exceptions specific to the Google Sheets adapter.
+"""
+
+from shillelagh.exceptions import Error
+
+
+class DateParseError(Error):
+    """
+    Raised when a date/time value cannot be parsed from a given pattern.
+    """
