+++ src/shillelagh/exceptions.py
@@ -13,7 +13,6 @@ __all__ = [
     "InternalError",
     "ProgrammingError",
     "NotSupportedError",
-    "DateParseError",
 ]
 
 
@@ -131,10 +130,4 @@ class ImpossibleFilterError(Error):
 class UnauthenticatedError(InterfaceError):
     """
     Custom excepton raised when the user needs to authenticate.
-    """
-
-
-class DateParseError(Error):
-    """
-    Raised when a date/time value cannot be parsed from a given pattern.
     """
